refactor: features/project flat 구조 재설계 및 네이밍 정리 (#261)
api/, hooks/, lib/ 중첩 폴더를 project.api.ts, project.hooks.ts,
project.lib.ts 로 통합하고 types.ts를 project.types.ts로 리네임.
queries/mutations 훅은 결정표에 따라 project.hooks.ts에 배치.
외부(app/mocks/review)의 deep import를 index.ts 경유로 정정.
