+++ app/(protected)/profile/[userId]/project/[projectId]/page.tsx
@@ -2,9 +2,7 @@ import { dehydrate } from '@tanstack/react-query';
 
 import { fetchMyProfile } from '@/features/profile/api/fetch-my-profile';
 import { profileKeys } from '@/features/profile/profile.query-keys';
-import { fetchProject } from '@/features/project/api/fetchers';
-import { projectKeys } from '@/features/project/api/query-keys';
-import { ProjectDetailStream } from '@/features/project/components/project-detail-stream';
+import { fetchProject, ProjectDetailStream, projectKeys } from '@/features/project';
 import { getQueryClient } from '@/shared/lib/get-query-client';
 
 interface Props {
