--- features/project/api/query-keys.ts
@@ -1,6 +1,0 @@
-export const projectKeys = {
-  all: ['project'] as const,
-  list: () => [...projectKeys.all, 'list'] as const,
-  detail: (projectId: number) => [...projectKeys.all, 'detail', projectId] as const,
-  byUser: (userId: number) => [...projectKeys.all, 'byUser', userId] as const,
-};
