+++ features/project/components/project-invite-handler.tsx
@@ -5,13 +5,12 @@ import { useCallback, useEffect, useMemo, useState } from 'react';
 
 import { saveLoginReturnPath, useLoginGate } from '@/features/auth';
 import { useMyProfile } from '@/features/profile/hooks/use-my-profile.query';
-import { useInviteProjectMember } from '@/features/project/api/mutations';
-import { useProject } from '@/features/project/api/queries';
 import { ProjectInviteAcceptModal } from '@/features/project/components/project-invite-accept-modal';
-import { parseProjectInviteId } from '@/features/project/lib/parse-project-invite-id';
-import { PROJECT_INVITE_QUERY_KEY } from '@/features/project/lib/project-invite-query';
 import { ConfirmationDialog } from '@/shared/components/dialog/confirmation-dialog';
 import { ApiError } from '@/shared/lib/api';
+
+import { useInviteProjectMember, useProject } from '../project.hooks';
+import { parseProjectInviteId, PROJECT_INVITE_QUERY_KEY } from '../project.lib';
 
 const removeInviteParamFromUrl = (pathname: string, searchParams: URLSearchParams): string => {
   const next = new URLSearchParams(searchParams.toString());
