+++ features/project/index.ts
@@ -1,15 +1,6 @@
 export { CreateProjectModal } from './components/create-project-modal';
 export { EditProjectModal } from './components/edit-project-modal';
 export { ProjectsPageContent } from './components/projects-page-content';
-export {
-  useConfirmProjectMembers,
-  useDeleteProject,
-  useDeleteProjectMember,
-  useInviteProjectMember,
-} from './api/mutations';
-export { fetchProjects } from './api/fetchers';
-export { useProject, useProjects, useUserProjects } from './api/queries';
-export { projectKeys } from './api/query-keys';
 export { ProjectPageContent } from './components/project-page-content';
 export { ProjectSection } from './components/project-section';
 export { ProjectStatusAlert } from './components/project-status-alert';
@@ -17,10 +8,29 @@ export { ProjectDetailStream } from './components/project-detail-stream';
 export { ProjectsStream } from './components/projects-stream';
 export { ProjectSectionStream } from './components/project-section-stream';
 export { ProjectSectionSkeleton } from './components/project-section.skeleton';
-export { useProjectCards } from './hooks/use-project-cards';
-export { useCreateProject } from './hooks/use-create-project.mutation';
-export { useUpdateProject } from './hooks/use-update-project.mutation';
-export { mapMyProjectToCardItem, mapMyProjectsToCardItems } from './lib/map-my-project-to-card';
+
+export { fetchProject, fetchProjects, projectKeys } from './project.api';
+
+export {
+  useConfirmProjectMembers,
+  useCreateProject,
+  useDeleteProject,
+  useDeleteProjectMember,
+  useInviteProjectMember,
+  useProject,
+  useProjectCards,
+  useProjects,
+  useUpdateProject,
+  useUserProjects,
+} from './project.hooks';
+
+export {
+  mapMyProjectsToCardItems,
+  mapMyProjectToCardItem,
+  parseProjectInviteId,
+  PROJECT_INVITE_QUERY_KEY,
+} from './project.lib';
+
 export type {
   MemberStatus,
   MyProjectResponse,
@@ -33,4 +43,4 @@ export type {
   ProjectResponse,
   ProjectStatus,
   UserProjectResponse,
-} from './types';
+} from './project.types';
