--- features/project/lib/parse-project-invite-id.ts
@@ -1,8 +1,0 @@
-export const parseProjectInviteId = (raw: string | null): number | null => {
-  if (raw == null || raw.trim() === '') {
-    return null;
-  }
-
-  const parsed = Number.parseInt(raw, 10);
-  return Number.isFinite(parsed) && parsed > 0 ? parsed : null;
-};
