--- features/project/lib/project-invite-query.ts
@@ -1,1 +1,0 @@
-export const PROJECT_INVITE_QUERY_KEY = 'inviteProjectId';
