+++ features/project/project.api.ts
@@ -1,5 +1,30 @@
-import type { MyProjectResponse, ProjectResponse, UserProjectResponse } from '@/features/project/types';
 import { apiRequest } from '@/shared/lib/api';
+
+import type {
+  MyProjectResponse,
+  ProjectPayload,
+  ProjectRequest,
+  ProjectResponse,
+  UserProjectResponse,
+} from './project.types';
+
+export const projectKeys = {
+  all: ['project'] as const,
+  list: () => [...projectKeys.all, 'list'] as const,
+  detail: (projectId: number) => [...projectKeys.all, 'detail', projectId] as const,
+  byUser: (userId: number) => [...projectKeys.all, 'byUser', userId] as const,
+};
+
+const createProjectFormData = (request: ProjectRequest, image?: File | null): FormData => {
+  const formData = new FormData();
+
+  formData.append('request', new Blob([JSON.stringify(request)], { type: 'application/json' }));
+  if (image) {
+    formData.append('image', image);
+  }
+
+  return formData;
+};
 
 /** GET /api/v1/projects 내 프로젝트 목록 */
 export const fetchProjects = (): Promise<MyProjectResponse[]> => apiRequest<MyProjectResponse[]>('/projects');
@@ -27,3 +52,17 @@ export const deleteProjectMember = (projectId: number, userId: number): Promise<
 /** PATCH /api/v1/projects/confirm/{projectId} 팀원 확정 */
 export const confirmProjectMembers = (projectId: number): Promise<void> =>
   apiRequest<void>(`/projects/confirm/${projectId}`, { method: 'PATCH' });
+
+/** POST /api/v1/projects 프로젝트 생성 */
+export const createProject = ({ request, image }: ProjectPayload): Promise<ProjectResponse> =>
+  apiRequest<ProjectResponse>('/projects', {
+    method: 'POST',
+    body: createProjectFormData(request, image),
+  });
+
+/** PATCH /api/v1/projects/{projectId} 프로젝트 수정 */
+export const updateProject = (projectId: number, { request, image }: ProjectPayload): Promise<ProjectResponse> =>
+  apiRequest<ProjectResponse>(`/projects/${projectId}`, {
+    method: 'PATCH',
+    body: createProjectFormData(request, image),
+  });
