+++ features/project/project.lib.ts
@@ -1,10 +1,69 @@
+import { SHARE_INVITER_NAME_PARAM } from '@/shared/constants/share-query';
+import { getSiteOrigin } from '@/shared/lib/get-site-origin';
+
 import type {
   MyProjectResponse,
   ProjectCardItem,
   ProjectCardMember,
   ProjectMemberResponse,
   ProjectMemberReviewStatus,
-} from '@/features/project/types';
+} from './project.types';
+import type { QueryClient } from '@tanstack/react-query';
+
+import { projectKeys } from './project.api';
+
+export const PROJECT_INVITE_QUERY_KEY = 'inviteProjectId';
+
+export const parseProjectInviteId = (raw: string | null): number | null => {
+  if (raw == null || raw.trim() === '') {
+    return null;
+  }
+
+  const parsed = Number.parseInt(raw, 10);
+  return Number.isFinite(parsed) && parsed > 0 ? parsed : null;
+};
+
+export const buildProjectInviteUrl = (projectId: number, inviterName?: string): string => {
+  const searchParams = new URLSearchParams({
+    [PROJECT_INVITE_QUERY_KEY]: String(projectId),
+  });
+
+  const trimmedName = inviterName?.trim();
+  if (trimmedName) {
+    searchParams.set(SHARE_INVITER_NAME_PARAM, trimmedName);
+  }
+
+  const path = `/projects?${searchParams.toString()}`;
+  const origin = getSiteOrigin();
+
+  return origin ? `${origin}${path}` : path;
+};
+
+type BuildReviewWriteUrlParams = {
+  projectId: number;
+  revieweeId: number;
+  revieweeName: string;
+};
+
+export const buildReviewWriteUrl = ({ projectId, revieweeId, revieweeName }: BuildReviewWriteUrlParams): string => {
+  const params = new URLSearchParams({
+    projectId: String(projectId),
+    revieweeId: String(revieweeId),
+    reviewee: revieweeName,
+  });
+
+  return `/reviews/new?${params.toString()}`;
+};
+
+/** 프로젝트 목록·프로필 탭 등 목록성 쿼리 갱신 */
+export const invalidateProjectListQueries = (queryClient: QueryClient, userId?: number) => {
+  void queryClient.invalidateQueries({ queryKey: projectKeys.list() });
+
+  if (userId != null && userId > 0) {
+    void queryClient.invalidateQueries({ queryKey: projectKeys.byUser(userId) });
+  }
+};
+
 type SessionInfo = { userId: number; nickname: string };
 
 const toReviewStatus = (member: ProjectMemberResponse, sessionUserId: number): ProjectMemberReviewStatus => {
